requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
ffmpeg-python==0.2.0
future==0.18.2
greenlet==2.0.1
idna==3.4
install==1.3.5
numpy==1.24.0
opencv-python==4.6.0.66
psycopg2==2.9.5
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.45
urllib3==1.26.13
